+++ CHANGELOG.md
@@ -26,6 +26,7 @@ with one *slight* but **important** difference:
   The following equality should always hold: `paths.reshape(*batch).shape = batch`.
 - Added the `differt.plugins` package and `differt.plugins.deepmimo` module (by <gh-user:jeertmans>, in <gh-pr:267>).
 - Added export utility to the DeepMIMO format (by <gh-user:jeertmans>, in <gh-pr:267>).
+- Added `from_mitsuba` and `from_sionna`methods to the `TriangleScene` class (by <gh-user:jeertmans>, in <gh-pr:267>).
 
 ### Chore